requirements.txt
asgiref==3.4.1
boto3==1.18.53
botocore==1.21.53
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
djangorestframework==3.12.4
idna==3.2
jmespath==0.10.0
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
cryptography==36.0.0